1. WayPoint.hpp - added Node as a type of WayPoint 2. NetworkLoader - added getter for the road network 3. RoadNetwork - added function getNodeById 4. SOCI_Converters - updated converting of turning group id 5. Removing unused functions

+++ dev/Basic/shared/geospatial/simmobility_network/SOCI_Converters.hpp
@@ -97,7 +97,7 @@ namespace soci
     typedef values base_type;
     static void from_base(const soci::values& vals, soci::indicator& ind, simmobility_network::TurningGroup& res)
     {
-      res.setTurningGroupId((unsigned int)vals.get<int>("id", 0));
+      res.setTurningGroupId(vals.get<unsigned int>("id", 0));
       res.setFromLinkId(vals.get<unsigned int>("from_link", 0));
       res.setNodeId(vals.get<unsigned int>("node_id", 0));
       res.setPhases(vals.get<std::string>("phases", ""));
